+++ changelog.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * Fixed `'<' not supported between instances of 'NoneType' and 'int'` error. Blank string for a decimal value is now converted to `None` when using the REST API.
+* Fix restoring table linking to trashed tables creating invalid link field. 
 
 ## Released (2022-01-13 1.8.2)